Refactor: 프로젝트 구조 개선을 위한 디렉토리 이동
